requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic>=0.40.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
